[frontend/backend] feat(threat-arsenal): review pr feedbacks (#4659)
